# openapi.yaml
openapi: 3.0.1
info:
  title: Learn New Lanaguages 
  description: A plugin that allows the user to learn german or french based on the CEFR Language units
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /setLanguage/{language}:
    post:
      operationId: setUserLanguage
      summary: Set the user's desired Language
      parameters:
            - in: path
              name: language
              schema:
                  type: string
              required: true
              description: German | French
      responses:
        '200':
          description: CEFR unit set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: The language was set to French
        '400':
          description: Invalid level provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid CEFR level
  /setUnit/{unitNumber}:
    post:
      operationId: setUnitGetPrompt
      summary: Set the user's desired CEFR level
      parameters:
            - in: path
              name: unitNumber
              schema:
                  type: string
              required: true
              description: the number of the unit
      responses:
        '200':
          description: CEFR unit set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: CEFR level set to Unit 1, this unit contains "How to correctly pronounce German words?"
        '400':
          description: Invalid level provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid CEFR level

  /teachUnit:
    get:
      operationId: teachUnit
      summary: Generate an easy to understand lesson about the unit content
      responses:
        '200':
          description: CEFR unit set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  gptPrompt:
                    type: string
                    example: You will act as a german teacher
                  unitContent:
                    type: string
                    example: Unit contains future tense verbs
        '400':
          description: something something meaningful error message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: something something meaningful error message

  /conversation:
    get:
      operationId: conversation
      summary: Talk to the user as a character of your choice to practice the language
      responses:
        '200':
          description: Roleplay as a native speaker of the target language
          content:
            application/json:
              schema:
                type: object
                properties:
                  gptPrompt:
                    type: string
                    example: You will roleplay as a german native speaker
                  unitContent:
                    type: string
                    example: Unit contains future tense verbs
        '400':
          description: something something meaningful error message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: something something meaningful error message